function getUserTransaction worked

+++ src/main/java/com/modsen/repository/entytie/UserTransactions.java
@@ -3,12 +3,13 @@ package com.modsen.repository.entytie;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 
 @Data
 @AllArgsConstructor
 public class UserTransactions {
-   private long id_beer;
+   private Integer id_beer;
    private int count;
    private Date date_time;
 }
